made factoring functional for all numbers between 2 and 100 and added formatting
The reason my program was running so slowly was that it was outside the range of an integer so there was a storage problem. To fix this, I did not calculate the factorial. Instead, I factored each number that makes the factorial. For example, 5! = 120 but instead of factoring 120, I factored 2, 3 ,4, and 5.